--- UpdateAbsoluteInFile/src/META-INF/services/com.agile.px.ICustomAction
@@ -1,1 +1,0 @@
-com.agile.px.UpdateAbsoluteInFile
