--- UpdateEffectivityDate/log4j.properties
@@ -1,16 +1,0 @@
-# Root logger option
-log4j.rootLogger=INFO, file, stdout
-	 
-# Direct log messages to stdout
-log4j.appender.stdout=org.apache.log4j.ConsoleAppender
-log4j.appender.stdout.Target=System.out
-log4j.appender.stdout.layout=org.apache.log4j.PatternLayout
-log4j.appender.stdout.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} [%t] %-5p:: %m%n
-
-# Direct log messages to a log file
-log4j.appender.file=org.apache.log4j.RollingFileAppender 
-log4j.appender.file.File=E:\\Agile\\PXLogs\\UpdateEffectivty.log
-log4j.appender.file.MaxFileSize=10000KB
-log4j.appender.file.MaxBackupIndex=10
-log4j.appender.file.layout=org.apache.log4j.PatternLayout
-log4j.appender.file.layout.ConversionPattern=%d{yyyy-MM-dd HH:mm:ss} [%t] %-5p:: %m%n
